split client_credential.py
