+++ sdk/identity/azure-identity/azure/identity/_credentials/certificate.py
@@ -1,0 +1,52 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+from typing import TYPE_CHECKING
+
+from .._authn_client import AuthnClient
+from .._base import CertificateCredentialBase
+
+if TYPE_CHECKING:
+    from azure.core.credentials import AccessToken
+    from typing import Any
+
+
+class CertificateCredential(CertificateCredentialBase):
+    """Authenticates as a service principal using a certificate.
+
+    :param str tenant_id: ID of the service principal's tenant. Also called its 'directory' ID.
+    :param str client_id: the service principal's client ID
+    :param str certificate_path: path to a PEM-encoded certificate file including the private key.
+
+    :keyword str authority: Authority of an Azure Active Directory endpoint, for example 'login.microsoftonline.com',
+          the authority for Azure Public Cloud (which is the default). :class:`~azure.identity.KnownAuthorities`
+          defines authorities for other clouds.
+    :keyword password: The certificate's password. If a unicode string, it will be encoded as UTF-8. If the certificate
+          requires a different encoding, pass appropriately encoded bytes instead.
+    :paramtype password: str or bytes
+    """
+
+    def get_token(self, *scopes, **kwargs):  # pylint:disable=unused-argument
+        # type: (*str, **Any) -> AccessToken
+        """Request an access token for `scopes`.
+
+        .. note:: This method is called by Azure SDK clients. It isn't intended for use in application code.
+
+        :param str scopes: desired scopes for the access token. This method requires at least one scope.
+        :rtype: :class:`azure.core.credentials.AccessToken`
+        :raises ~azure.core.exceptions.ClientAuthenticationError: authentication failed. The error's ``message``
+          attribute gives a reason. Any error response from Azure Active Directory is available as the error's
+          ``response`` attribute.
+        """
+        if not scopes:
+            raise ValueError("'get_token' requires at least one scope")
+
+        token = self._client.get_cached_token(scopes)
+        if not token:
+            data = self._get_request_data(*scopes)
+            token = self._client.request_token(scopes, form_data=data)
+        return token
+
+    def _get_auth_client(self, tenant_id, **kwargs):
+        return AuthnClient(tenant=tenant_id, **kwargs)
