+++ sdk/identity/azure-identity/azure/identity/_credentials/environment.py
@@ -6,7 +6,8 @@ import os
 
 from .. import CredentialUnavailableError
 from .._constants import EnvironmentVariables
-from .client_credential import CertificateCredential, ClientSecretCredential
+from .certificate import CertificateCredential
+from .client_secret import ClientSecretCredential
 from .user_password import UsernamePasswordCredential
