+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/certificate.py
@@ -6,57 +6,11 @@ from typing import TYPE_CHECKING
 
 from .base import AsyncCredentialBase
 from .._authn_client import AsyncAuthnClient
-from ..._base import ClientSecretCredentialBase, CertificateCredentialBase
+from ..._base import CertificateCredentialBase
 
 if TYPE_CHECKING:
-    from typing import Any, Mapping
+    from typing import Any
     from azure.core.credentials import AccessToken
-
-
-class ClientSecretCredential(ClientSecretCredentialBase, AsyncCredentialBase):
-    """Authenticates as a service principal using a client ID and client secret.
-
-    :param str tenant_id: ID of the service principal's tenant. Also called its 'directory' ID.
-    :param str client_id: the service principal's client ID
-    :param str client_secret: one of the service principal's client secrets
-
-    :keyword str authority: Authority of an Azure Active Directory endpoint, for example 'login.microsoftonline.com',
-          the authority for Azure Public Cloud (which is the default). :class:`~azure.identity.KnownAuthorities`
-          defines authorities for other clouds.
-    """
-
-    def __init__(self, tenant_id: str, client_id: str, client_secret: str, **kwargs: "Any") -> None:
-        super(ClientSecretCredential, self).__init__(tenant_id, client_id, client_secret, **kwargs)
-        self._client = AsyncAuthnClient(tenant=tenant_id, **kwargs)
-
-    async def __aenter__(self):
-        await self._client.__aenter__()
-        return self
-
-    async def close(self):
-        """Close the credential's transport session."""
-
-        await self._client.__aexit__()
-
-    async def get_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":  # pylint:disable=unused-argument
-        """Asynchronously request an access token for `scopes`.
-
-        .. note:: This method is called by Azure SDK clients. It isn't intended for use in application code.
-
-        :param str scopes: desired scopes for the access token. This method requires at least one scope.
-        :rtype: :class:`azure.core.credentials.AccessToken`
-        :raises ~azure.core.exceptions.ClientAuthenticationError: authentication failed. The error's ``message``
-          attribute gives a reason. Any error response from Azure Active Directory is available as the error's
-          ``response`` attribute.
-        """
-        if not scopes:
-            raise ValueError("'get_token' requires at least one scope")
-
-        token = self._client.get_cached_token(scopes)
-        if not token:
-            data = dict(self._form_data, scope=" ".join(scopes))
-            token = await self._client.request_token(scopes, form_data=data)
-        return token  # type: ignore
 
 
 class CertificateCredential(CertificateCredentialBase, AsyncCredentialBase):
